Задание 9
